+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Made number of events per day in month view dynamics - [#323](https://github.com/hkdb/aerion/issues/323)
 - Fixed icon rendering with newer DEs and compositors - [#316](https://github.com/hkdb/aerion/issues/316)
 - Fixed diff tz time display in calendar
+- Fixed calendar about field plaintext rendering
 
 
 **v0.3.0 - 06-23-2026**